train/requirements.txt
numpy
torch
python-chess
tqdm